Fixed hardcoded value and project dependency.
xcopy had a debug value hardcoded this makes the build fail on release
since xcopy fails.
If the intention is to run only on debug this should be guarded with a
condition

Added a dependency on OctoFx.Tests for OctoFX.TradingWebsite project,
since there was no dependency specified, build can fail when msbuild is
used (msbuild doesn't guarantees ordering) if OctoFX.Tests have not
build (xcopy fails).